--- src/main/java/java/pro/sky/java/course2/coursework2examiner/CourseWork2EximinerApplication.java
@@ -1,13 +1,0 @@
-package java.pro.sky.java.course2.coursework2examiner;
-
-import org.springframework.boot.SpringApplication;
-import org.springframework.boot.autoconfigure.SpringBootApplication;
-
-@SpringBootApplication
-public class CourseWork2EximinerApplication {
-
-    public static void main(String[] args) {
-        SpringApplication.run(CourseWork2EximinerApplication.class, args);
-    }
-
-}
