--- src/main/java/java/pro/sky/java/course2/coursework2examiner/controllers/QuestionController.java
@@ -1,35 +1,0 @@
-package java.pro.sky.java.course2.coursework2examiner.controllers;
-
-import org.springframework.web.bind.annotation.GetMapping;
-import org.springframework.web.bind.annotation.RequestMapping;
-import org.springframework.web.bind.annotation.RequestParam;
-import org.springframework.web.bind.annotation.RestController;
-
-import java.pro.sky.java.course2.coursework2examiner.model.Question;
-import java.pro.sky.java.course2.coursework2examiner.service.QuestionService;
-import java.util.Collection;
-
-@RestController
-@RequestMapping("/exam/java")
-public class QuestionController {
-    final private QuestionService questionService;
-
-    public QuestionController(QuestionService questionService) {
-        this.questionService = questionService;
-    }
-
-    @GetMapping("/add")
-    public Question addQuestion(@RequestParam String question, @RequestParam String answer) {
-        return questionService.addQuestion(new Question(question, answer));
-    }
-
-    @GetMapping("/remove")
-    public Question removeQuestion(@RequestParam String question, @RequestParam String answer) {
-        return questionService.removeQuestion(new Question(question, answer));
-    }
-
-    @GetMapping
-    public Collection<Question> getAllQuestions() {
-        return questionService.getAllQuestions();
-    }
-}
