--- src/main/java/java/pro/sky/java/course2/coursework2examiner/exceptions/AmountBiggerThanQuestionsSet.java
@@ -1,22 +1,0 @@
-package java.pro.sky.java.course2.coursework2examiner.exceptions;
-
-import org.springframework.http.HttpStatus;
-import org.springframework.web.bind.annotation.ResponseStatus;
-
-@ResponseStatus(HttpStatus.BAD_REQUEST)
-public class AmountBiggerThanQuestionsSet extends RuntimeException{
-    public AmountBiggerThanQuestionsSet() {
-    }
-
-    public AmountBiggerThanQuestionsSet(String message) {
-        super(message);
-    }
-
-    public AmountBiggerThanQuestionsSet(String message, Throwable cause) {
-        super(message, cause);
-    }
-
-    public AmountBiggerThanQuestionsSet(Throwable cause) {
-        super(cause);
-    }
-}
