--- src/main/java/java/pro/sky/java/course2/coursework2examiner/exceptions/QuestionNotExistException.java
@@ -1,22 +1,0 @@
-package java.pro.sky.java.course2.coursework2examiner.exceptions;
-
-import org.springframework.http.HttpStatus;
-import org.springframework.web.bind.annotation.ResponseStatus;
-
-@ResponseStatus(HttpStatus.NOT_FOUND)
-public class QuestionNotExistException extends RuntimeException {
-    public QuestionNotExistException() {
-    }
-
-    public QuestionNotExistException(String message) {
-        super(message);
-    }
-
-    public QuestionNotExistException(String message, Throwable cause) {
-        super(message, cause);
-    }
-
-    public QuestionNotExistException(Throwable cause) {
-        super(cause);
-    }
-}
