--- src/main/java/java/pro/sky/java/course2/coursework2examiner/service/ExaminerServiceImpl.java
@@ -1,27 +1,0 @@
-package java.pro.sky.java.course2.coursework2examiner.service;
-
-import org.springframework.stereotype.Service;
-
-import java.pro.sky.java.course2.coursework2examiner.exceptions.AmountBiggerThanQuestionsSet;
-import java.pro.sky.java.course2.coursework2examiner.model.Question;
-import java.util.HashSet;
-import java.util.Set;
-@Service
-public class ExaminerServiceImpl implements ExaminerService {
-    private final QuestionService questionService;
-
-    public ExaminerServiceImpl(QuestionService questionService) {
-        this.questionService = questionService;
-    }
-    @Override
-    public Set<Question> getQuastions(Integer amount) {
-        if (amount > questionService.getAllQuestions().size()) {
-            throw new AmountBiggerThanQuestionsSet();
-        }
-        Set<Question> questionSet = new HashSet<>();
-        while (questionSet.size() < amount) {
-            questionSet.add(questionService.getRandomQuestion());
-        }
-        return questionSet;
-    }
-}
