+++ src/main/java/java2/pro/sky/java/course2/coursework2examiner/model/Question.java
@@ -1,0 +1,35 @@
+package java2.pro.sky.java.course2.coursework2examiner.model;
+
+public class Question {
+    private String question;
+    private String answer;
+
+    public Question(String question, String answer) {
+        this.question = question;
+        this.answer = answer;
+    }
+
+    public String getQuestion() {
+        return question;
+    }
+
+    public void setQuestion(String question) {
+        this.question = question;
+    }
+
+    public String getAnswer() {
+        return answer;
+    }
+
+    public void setAnswer(String answer) {
+        this.answer = answer;
+    }
+
+    @Override
+    public String toString() {
+        return "Question{" +
+                "question='" + question + '\'' +
+                ", anwser='" + answer + '\'' +
+                '}';
+    }
+}
