+++ src/test/java/java2/pro/sky/java/course2/coursework2examiner/CourseWork2EximinerApplicationTests.java
@@ -1,0 +1,13 @@
+package java2.pro.sky.java.course2.coursework2examiner;
+
+import org.junit.jupiter.api.Test;
+import org.springframework.boot.test.context.SpringBootTest;
+
+@SpringBootTest
+class CourseWork2EximinerApplicationTests {
+
+    @Test
+    void contextLoads() {
+    }
+
+}
